Lab3/requirements.txt
grpcio
grpcio-tools
requests
readerwriterlock
pandas